+++ DataLayers/Framework/ILayer.cs
@@ -23,7 +23,7 @@ namespace Pathoschild.Stardew.DataLayers.Framework
         LegendEntry[] Legend { get; }
 
         /// <summary>The buttons to activate the layer.</summary>
-        SButton[] LayerButtons { get; }
+        SButton[] ShortcutKey { get; }
 
 
         /*********
@@ -35,9 +35,5 @@ namespace Pathoschild.Stardew.DataLayers.Framework
         /// <param name="visibleTiles">The tile positions currently visible on the screen.</param>
         /// <param name="cursorTile">The tile position under the cursor.</param>
         TileGroup[] Update(GameLocation location, in Rectangle visibleArea, in Vector2[] visibleTiles, in Vector2 cursorTile);
-
-        /// <summary>Parses the shortcuts from the config.</summary>
-        /// <param name="monitor">The monitor through which to log an error if a button value is invalid.</param>
-        void ParseShortcuts(IMonitor monitor);
     }
 }