+++ DataLayers/ModEntry.cs
@@ -30,6 +30,9 @@ namespace Pathoschild.Stardew.DataLayers
         /// <summary>The available data layers.</summary>
         private ILayer[] Layers;
 
+        /// <summary>Maps button presses to the layer it should activate.</summary>
+        private readonly IDictionary<SButton, ILayer> ShortcutMap = new Dictionary<SButton, ILayer>();
+
         /// <summary>Handles access to the supported mod integrations.</summary>
         private ModIntegrations Mods;
 
@@ -69,9 +72,14 @@ namespace Pathoschild.Stardew.DataLayers
             this.Mods = new ModIntegrations(this.Monitor, this.Helper.ModRegistry, this.Helper.Reflection);
             this.Layers = this.GetLayers(this.Config, this.Helper.Translation, this.Mods).ToArray();
 
+            // cache shortcut keys
             foreach (ILayer layer in this.Layers)
             {
-                layer.ParseShortcuts(this.Monitor);
+                foreach (SButton button in layer.ShortcutKey)
+                {
+                    if (!this.ShortcutMap.ContainsKey(button))
+                        this.ShortcutMap[button] = layer;
+                }
             }
         }
 
@@ -84,29 +92,29 @@ namespace Pathoschild.Stardew.DataLayers
             ModConfig.LayerConfigs layers = config.Layers;
 
             if (layers.Accessible.IsEnabled())
-                yield return new AccessibleLayer(translation, layers.Accessible);
+                yield return new AccessibleLayer(translation, layers.Accessible, this.Monitor);
             if (layers.Buildable.IsEnabled())
-                yield return new BuildableLayer(translation, layers.Buildable);
+                yield return new BuildableLayer(translation, layers.Buildable, this.Monitor);
             if (layers.CoverageForBeeHouses.IsEnabled())
-                yield return new BeeHouseLayer(translation, layers.CoverageForBeeHouses);
+                yield return new BeeHouseLayer(translation, layers.CoverageForBeeHouses, this.Monitor);
             if (layers.CoverageForScarecrows.IsEnabled())
-                yield return new ScarecrowLayer(translation, layers.CoverageForScarecrows, mods);
+                yield return new ScarecrowLayer(translation, layers.CoverageForScarecrows, mods, this.Monitor);
             if (layers.CoverageForSprinklers.IsEnabled())
-                yield return new SprinklerLayer(translation, layers.CoverageForSprinklers, mods);
+                yield return new SprinklerLayer(translation, layers.CoverageForSprinklers, mods, this.Monitor);
             if (layers.CoverageForJunimoHuts.IsEnabled())
-                yield return new JunimoHutLayer(translation, layers.CoverageForJunimoHuts, mods);
+                yield return new JunimoHutLayer(translation, layers.CoverageForJunimoHuts, mods, this.Monitor);
             if (layers.CropWater.IsEnabled())
-                yield return new CropWaterLayer(translation, layers.CropWater);
+                yield return new CropWaterLayer(translation, layers.CropWater, this.Monitor);
             if (layers.CropPaddyWater.IsEnabled())
-                yield return new CropPaddyWaterLayer(translation, layers.CropPaddyWater);
+                yield return new CropPaddyWaterLayer(translation, layers.CropPaddyWater, this.Monitor);
             if (layers.CropFertilizer.IsEnabled())
-                yield return new CropFertilizerLayer(translation, layers.CropFertilizer);
+                yield return new CropFertilizerLayer(translation, layers.CropFertilizer, this.Monitor);
             if (layers.CropHarvest.IsEnabled())
-                yield return new CropHarvestLayer(translation, layers.CropHarvest);
+                yield return new CropHarvestLayer(translation, layers.CropHarvest, this.Monitor);
             if (layers.Machines.IsEnabled() && mods.Automate.IsLoaded)
-                yield return new MachineLayer(translation, layers.Machines, mods);
+                yield return new MachineLayer(translation, layers.Machines, mods, this.Monitor);
             if (layers.Tillable.IsEnabled())
-                yield return new TillableLayer(translation, layers.Tillable);
+                yield return new TillableLayer(translation, layers.Tillable, this.Monitor);
         }
 
         /// <summary>The method invoked when the player returns to the title screen.</summary>
@@ -156,53 +164,14 @@ namespace Pathoschild.Stardew.DataLayers
                     this.CurrentOverlay.PrevLayer();
                     this.Helper.Input.Suppress(e.Button);
                 }
-                else
+
+                // shortcut to layer
+                else if (overlayVisible && this.ShortcutMap.TryGetValue(e.Button, out ILayer layer) && layer != this.CurrentOverlay.CurrentLayer)
                 {
-
-                    ILayer layerForButton = this.GetLayerForButton(e.Button);
-
-                    if (layerForButton != null)
-                    {
-                        if (this.Config.ActivateOverlayOnShortcut && !overlayVisible)
-                        {
-                            this.CurrentOverlay = new DataLayerOverlay(this.Helper.Events, this.Helper.Input, this.Layers, this.CanOverlayNow, this.Config.CombineOverlappingBorders, this.Config.ShowGrid);
-                            overlayVisible = this.CurrentOverlay != null;
-                        }
-
-                        if (overlayVisible)
-                        {
-                                if (this.CurrentOverlay.CurrentLayer == layerForButton && this.Config.ToggleOverlayViaShortcut)
-                                {
-                                    this.CurrentOverlay.Dispose();
-                                    this.CurrentOverlay = null;
-                                    this.Helper.Input.Suppress(e.Button);
-                                }
-                                else
-                                {
-                                    this.CurrentOverlay.ActivateLayer(layerForButton);
-                                    this.Helper.Input.Suppress(e.Button);
-                                }
-
-                        }
-                    }
+                    this.CurrentOverlay.SetLayer(layer);
+                    this.Helper.Input.Suppress(e.Button);
                 }
             });
-        }
-
-        /// <summary>Retrieves a layer by it's configured shortcut button.</summary>
-        /// <param name="button">The button the game received.</param>
-        /// <returns>The found layer or null.</returns>
-        private ILayer GetLayerForButton(SButton button)
-        {
-            foreach (var layer in this.Layers)
-            {
-                if (layer.LayerButtons != null && layer.LayerButtons.Contains(button))
-                {
-                    return layer;
-                }
-            }
-
-            return null;
         }
 
         /// <summary>Receive an update tick.</summary>